+++ src/Controller/UserFamilyController.php
@@ -11,8 +11,11 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/user_family')]
+#[IsGranted('ROLE_USER')]
+
 class UserFamilyController extends AbstractController
 {
     private $security;
@@ -57,6 +60,10 @@ class UserFamilyController extends AbstractController
     #[Route('/{id}', name: 'app_user_family_show', methods: ['GET'])]
     public function show(UserFamily $userFamily): Response
     {
+        if ($userFamily->getUser() != $this->getUser()) {
+            return  $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('user_family/show.html.twig', [
             'user_family' => $userFamily,
         ]);
@@ -65,6 +72,9 @@ class UserFamilyController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_family_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, UserFamily $userFamily, EntityManagerInterface $entityManager): Response
     {
+        if ($userFamily->getUser() != $this->getUser()) {
+            return  $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(UserFamilyType::class, $userFamily);
         $form->handleRequest($request);
 
@@ -83,7 +93,11 @@ class UserFamilyController extends AbstractController
     #[Route('/{id}', name: 'app_user_family_delete', methods: ['POST'])]
     public function delete(Request $request, UserFamily $userFamily, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$userFamily->getId(), $request->request->get('_token'))) {
+        if ($userFamily->getUser() != $this->getUser()) {
+            return  $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
+        
+        if ($this->isCsrfTokenValid('delete' . $userFamily->getId(), $request->request->get('_token'))) {
             $entityManager->remove($userFamily);
             $entityManager->flush();
         }
